requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlmodel>=0.0.14
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.1.0
pydantic[email]>=2.5.0
alembic>=1.13.0
motor
beanie
pytest>=7.4.3
pytest-asyncio>=0.23.2
pytest-cov
slowapi
python-dotenv
fastapi-limiter
redis
httpx>=0.25.2
PyJWT
fakeredis
pymongo==4.13.2